Move debugger type proxies in the same namespace as their types.
